refactor: move utility files from composables to utils folder
Move non-composable utility files to their proper location:
- debug-utils.js → app/utils/
- promise-cache-utils.js → app/utils/

These files are pure utility functions without Vue reactivity,
so they belong in utils/ rather than composables/.

Updated all imports in:
- app/stores/audioStore.js
- app/stores/userStore.js
- app/stores/voiceStore.js
- app/utils/frequency-analyzer.js
- 4 test files
